// src/interfaces/comlink/playerData.ts
export interface PlayerData {
  rosterUnit: [];
  playerRating: [];
  profileStat: [];
  pvpProfile: [];
  unlockedPlayerTitle: [];
  unlockedPlayerPortrait: [];
  seasonStatus: [];
  datacron: [];
  selectedPlayerTitle: object;
  selectedPlayerPortrait: object;
  name: string;
  level: number;
  allyCode: string;
  playerId: string;
  guildId: string;
  guildName: string;
  guildLogoBackground: string;
  guildBannerColor: string;
  guildBannerLogo: string;
  guildTypeId: string;
  localTimeZoneOffsetMinutes: number;
  lastActivityTime: Date;
  lifetimeSeasonScore: string;
}